Refactor iOS progress store
